+++ Persistence/src/main/java/com/kalaazu/persistence/service/LevelsService.java
@@ -4,11 +4,11 @@ import com.kalaazu.persistence.entity.LevelsEntity;
 
 /**
  * Levels service.
- * ==================
+ * ===============
  *
  * Service for the Levels entity.
  *
  * @author Manulaiko <manulaiko@gmail.com>
  */
-public interface LevelsService extends IService<LevelsEntity, Integer> {
+public interface LevelsService extends IService<LevelsEntity, Byte> {
 }
