+++ Persistence/src/main/java/com/kalaazu/persistence/service/MapsStationsServiceImpl.java
@@ -32,7 +32,7 @@ public class MapsStationsServiceImpl implements MapsStationsService {
      * @inheritDoc
      */
     @Override
-    public MapsStationsEntity find(Integer id) {
+    public MapsStationsEntity find(Byte id) {
         return this.repository.findById(id).orElse(null);
     }
 
@@ -56,7 +56,7 @@ public class MapsStationsServiceImpl implements MapsStationsService {
      * @inheritDoc
      */
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(Byte id) {
         this.repository.deleteById(id);
 
         return !this.repository.existsById(id);
